\name{DW}
\alias{DW}
\docType{data}
\title{Dehejia-Wahba dataset}
\description{
A subset of the Lalonde dataset (see cited reference).}
\usage{data(DW)}
\format{
  A data frame with 445 observations on the following 10 variables.
  \describe{
    \item{\code{treated}}{treated variable indicator}
    \item{\code{age}}{age}
    \item{\code{education}}{years of education}
    \item{\code{black}}{race indicator variable}
    \item{\code{married}}{marital status indicator variable}
    \item{\code{nodegree}}{indicator variable of not possessing a degree}
    \item{\code{re74}}{real earnings in 1974}
    \item{\code{re75}}{real earnings in 1975}
    \item{\code{re78}}{real earnings in 1978 (post treatment outcome)}
    \item{\code{hispanic}}{ethnic indicator variable}
    \item{\code{u74}}{unemployment in 1974 indicator variable}
    \item{\code{u75}}{unemployment in 1975 indicator variable}
  }
}
\source{see references}
\references{
Dehejia, R., Wahba, S. (1999) ``Causal Effects in Nonexperimental Studies: 
Reevaluating the Evaluation of Training Programs,'' \emph{Journal of the American 
Statistical Association}, 94, 1053-1062.
}
\keyword{datasets}
