\name{LL}
\alias{LL}
\docType{data}
\title{Lalonde dataset}
\description{
Lalonde experimental dataset  (see cited reference).
}
\usage{data(LL)}
\format{
  A data frame with 722 observations on the following 10 variables.
  \describe{
    \item{\code{treated}}{treatment variable indicator}
    \item{\code{age}}{age}
    \item{\code{education}}{years of education}
    \item{\code{black}}{race indicator variable}
    \item{\code{married}}{marital status indicator variable}
    \item{\code{nodegree}}{indicator variable for not possessing a degree}
    \item{\code{re74}}{real earnings in 1974}
    \item{\code{re75}}{real earnings in 1975}
    \item{\code{re78}}{real earnings in 1978 (post-treatment outcome)}
    \item{\code{hispanic}}{ethnic indicator variable}
    \item{\code{u74}}{unemployment in 1974 indicator variable}
    \item{\code{u75}}{unemployment in 1975 indicator variable}
  }
}
\source{see references}
\references{
Lalonde (1986) Evaluating the Econometric Evaluations of Training Programs with Experimental Data, \emph{The American Economic Review},
76(4), 604-620. 
}
\keyword{datasets}
